[AoC] Day 20 Part 2
This is miserable. I had to analyze the graph nodes to figure out what the circuit was actually doing and decide that the NANDs were the main cycle. Their outputs link into rx, so it's only once their cycles align that rx lights up.
But this isn't a generic solution. It assumes explicitly that rx has only nands leading into it.